Fix QA issues: half-star ratings, bag chip clear, rating sheet scroll
- Change BrewLogEntity.rating from Int to Float for half-star precision
- Update StarRatingRow to render half-star icons (StarHalf)
- Fix bag chip: tap clears selection, shows Close icon, re-tap opens picker
- Add verticalScroll to BrewRatingSheet for small screens
- Bump DB version 3→4 for schema change
- Fix test assertions for Float rating

+++ app/src/main/java/com/adsamcik/starlitcoffee/data/db/dao/BrewLogDao.kt
@@ -13,7 +13,7 @@ interface BrewLogDao {
     suspend fun insert(log: BrewLogEntity): Long
 
     @Query("UPDATE brew_logs SET rating = :rating, freeformNotes = :notes WHERE id = :logId")
-    suspend fun updateRating(logId: Long, rating: Int, notes: String?)
+    suspend fun updateRating(logId: Long, rating: Float, notes: String?)
 
     @Query("SELECT * FROM brew_logs ORDER BY createdAt DESC")
     fun getAll(): Flow<List<BrewLogEntity>>
